批量专用逻辑搬进 BatchPatternUploadUtil，缩小共享工具类的改动面
ExtendedAEPatternUploadUtil 已有两千行，是单个上传路径与本 PR 共用的类；
往里塞三百行只有批量路径会用的代码，评审时很难分辨哪些改动会影响既有行为。
改成新建 BatchPatternUploadUtil 承载批量独有逻辑，共享类里只留必须动的部分：

- 查重整块（样板规范化、哈希集合、collectExistingPatterns 等）搬到新类
- uploadPatternToMatchingProvider 与 providerNameMatches、listAllProvidersFromGrid 一并搬走
- matrixContainsPattern 恢复原实现：先前把它改成调用新的 isSamePattern，
  虽然等价（空栈判定与默认 CompoundTag 两侧行为一致），但那是重写既有代码，没必要
- 顺带删掉三个没有调用方的方法：matrixHasPattern、providerHasPattern、isSamePattern
- 映射查询（getMappingVersion / getRecipeTypeId / mappedSearchKeyOrNull 等）只有
  客户端合成树标记会用，本 PR 里没有调用方，挪到用得上它的那个 PR

共享类里剩下的改动只有两类：uploadPatternToMatrix 的 quiet 参数（3 参数版原样保留
并以 quiet=false 委托，既有调用方输出不变），以及 findAllMatrixPatternInventories
和 MatrixInventoryTarget 由 private 放宽到包内可见，供新类扫描装配矩阵。
共 25 行新增、10 行改动，既有方法实现零重写。
