+++ src/main/java/com/extendedae_plus/network/BatchCreateAndUploadPatternC2SPacket.java
@@ -4,6 +4,7 @@ import appeng.api.networking.IGrid;
 import appeng.helpers.patternprovider.PatternContainer;
 import com.extendedae_plus.ExtendedAEPlus;
 import com.extendedae_plus.util.uploadPattern.CtrlQPendingUploadUtil;
+import com.extendedae_plus.util.uploadPattern.BatchPatternUploadUtil;
 import com.extendedae_plus.util.uploadPattern.ExtendedAEPatternUploadUtil;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.chat.Component;
@@ -108,7 +109,7 @@ public class BatchCreateAndUploadPatternC2SPacket implements CustomPacketPayload
 			// 供应器列表整批复用一次快照：逐个查会在大树上重复遍历整个网络的机器类。
 			List<PatternContainer> providers = CtrlQPendingUploadUtil.listAvailableProvidersFromPlayerNetwork(player);
 			// 已存在样板集合同样只建一次：逐个样板扫全网是 O(样板数 × 供应器数 × 槽位数)。
-			Set<ItemStack> existingPatterns = ExtendedAEPatternUploadUtil.collectExistingPatterns(grid, providers);
+			Set<ItemStack> existingPatterns = BatchPatternUploadUtil.collectExistingPatterns(grid, providers);
 
 			int toMatrix = 0;
 			int toProvider = 0;
@@ -141,7 +142,7 @@ public class BatchCreateAndUploadPatternC2SPacket implements CustomPacketPayload
 				}
 
 				// 反复按一键编码时不该在网络里堆一摞重复样板：装配矩阵与供应器都算已存在。
-				if (ExtendedAEPatternUploadUtil.containsPattern(existingPatterns, pattern)) {
+				if (BatchPatternUploadUtil.containsPattern(existingPatterns, pattern)) {
 					duplicate++;
 					continue;
 				}
@@ -153,21 +154,21 @@ public class BatchCreateAndUploadPatternC2SPacket implements CustomPacketPayload
 
 				// 装配矩阵只收合成/锻造/切石样板，处理样板必然落到下一段。
 				if (ExtendedAEPatternUploadUtil.uploadPatternToMatrix(player, pattern, grid, true)) {
-					ExtendedAEPatternUploadUtil.rememberPattern(existingPatterns, pattern);
+					BatchPatternUploadUtil.rememberPattern(existingPatterns, pattern);
 					toMatrix++;
 					continue;
 				}
 
-				if (ExtendedAEPatternUploadUtil.uploadPatternToMatchingProvider(
+				if (BatchPatternUploadUtil.uploadPatternToMatchingProvider(
 						player, pattern, providers, entry.providerSearchKey())) {
-					ExtendedAEPatternUploadUtil.rememberPattern(existingPatterns, pattern);
+					BatchPatternUploadUtil.rememberPattern(existingPatterns, pattern);
 					toProvider++;
 					continue;
 				}
 
 				// 落背包的也登记：服务端不能依赖客户端已按配方去重，
 				// 否则一个塞满同一条目的封包会白扣掉整叠空白样板。
-				ExtendedAEPatternUploadUtil.rememberPattern(existingPatterns, pattern);
+				BatchPatternUploadUtil.rememberPattern(existingPatterns, pattern);
 				if (!player.getInventory().add(pattern)) {
 					player.drop(pattern.copy(), false);
 				}
